Some general improvements/fixes

+++ Server/src/org/crandor/game/node/item/WeightedChanceItem.java
@@ -4,16 +4,16 @@ import org.crandor.tools.RandomFunction;
 
 public class WeightedChanceItem {
     int id,minimum_amount,maximum_amount;
-    public int weight;
+    public double weight;
 
-    public WeightedChanceItem(int id, int minimum_amount, int maximum_amount, int weight){
+    public WeightedChanceItem(int id, int minimum_amount, int maximum_amount, double weight){
         this.id = id;
         this.minimum_amount = minimum_amount;
         this.maximum_amount = maximum_amount;
         this.weight = weight;
     }
 
-    public WeightedChanceItem(int id, int amount, int weight){
+    public WeightedChanceItem(int id, int amount, double weight){
         this(id,amount,amount,weight);
     }
